[scheduler] Split the dependency arrow geometry module
Pure move: the anchor resolver, the orthogonal routing and the hit-area helpers
get a module each, so the routing changes for the new dependency types are
reviewable on their own.
