+++ packages/x-scheduler-premium/src/event-timeline-premium/content/timeline-dependency-arrows/dependencyArrowRouting.ts
@@ -1,0 +1,196 @@
+import type { DependencyArrowObstacle, DependencyArrowPoint } from './dependencyAnchorResolver';
+
+/**
+ * Minimum horizontal segment when leaving the predecessor's end edge and when
+ * entering the successor's start edge.
+ */
+const DEPENDENCY_ARROW_STUB = 8;
+/**
+ * Radius used to soften the corners of the orthogonal route.
+ */
+export const DEPENDENCY_ARROW_CORNER_RADIUS = 4;
+/**
+ * Size of the arrowhead marker at the successor's start edge.
+ */
+export const DEPENDENCY_ARROWHEAD_SIZE = 7;
+/**
+ * Minimum length of the final segment entering the target: the softened corner plus the
+ * arrowhead must fit on it, otherwise the arrowhead overlaps the curve.
+ */
+const DEPENDENCY_ARROW_TARGET_CLEARANCE =
+  DEPENDENCY_ARROW_CORNER_RADIUS + DEPENDENCY_ARROWHEAD_SIZE + 1;
+
+/**
+ * Builds the candidate orthogonal routes from the source anchor (end edge of the
+ * predecessor) to the target anchor (start edge of the successor), best first.
+ * The forward elbow returns two candidates (turn right after the source, or right
+ * before the target) so the caller can pick the one crossing the fewest events.
+ * `detourOffset` is how far from the source anchor the S route runs its horizontal
+ * detour — it must clear the event's edge, otherwise the route overlaps the events and
+ * reads as a knot instead of a detour.
+ * Routes stay inside `[0, eventsWidth]`: at a timeline edge the stubs ride over the
+ * event instead of leaving the visible area.
+ */
+export function buildDependencyArrowRoutes(
+  source: DependencyArrowPoint,
+  target: DependencyArrowPoint,
+  detourOffset: number,
+  eventsWidth: number,
+): DependencyArrowPoint[][] {
+  const forwardX = target.x - source.x;
+
+  if (source.y === target.y && forwardX >= 0) {
+    if (forwardX >= 2 * DEPENDENCY_ARROW_STUB) {
+      // Straight segment between two anchors at the same height.
+      return [[source, target]];
+    }
+
+    // Adjacent events (two same-lane events never overlap in time, so a same-height
+    // gap is always forward): a short straight arrow slightly overlapping the
+    // predecessor's tail reads better than a detour around the junction.
+    return [[{ x: Math.max(0, target.x - 2 * DEPENDENCY_ARROW_STUB), y: target.y }, target]];
+  }
+
+  if (forwardX >= DEPENDENCY_ARROW_STUB + DEPENDENCY_ARROW_TARGET_CLEARANCE) {
+    // Forward elbow: right off the source, vertical, then right into the target — or
+    // the mirrored turn right before the target.
+    const earlyTurnX = source.x + DEPENDENCY_ARROW_STUB;
+    const lateTurnX = target.x - DEPENDENCY_ARROW_TARGET_CLEARANCE;
+    const routes = [
+      [source, { x: earlyTurnX, y: source.y }, { x: earlyTurnX, y: target.y }, target],
+    ];
+    if (lateTurnX > earlyTurnX) {
+      routes.push([source, { x: lateTurnX, y: source.y }, { x: lateTurnX, y: target.y }, target]);
+    }
+    return routes;
+  }
+
+  // S route: the successor starts before (or too close to) the predecessor's end, so
+  // the arrow exits right, detours horizontally hugging the source event (below it, or
+  // above when the target is higher) and comes back before entering the target. An
+  // arbitrary height between the two anchors could land exactly on a row border and
+  // read as part of the grid.
+  // The verticals clamp to the events area (x < 0 sits under the pinned title column)
+  // and the fixed-length stubs then ride over the event — the same overlap trade-off
+  // as the short arrow between two adjacent events.
+  const detourY = target.y >= source.y ? source.y + detourOffset : source.y - detourOffset;
+  const exitX = Math.min(eventsWidth, source.x + DEPENDENCY_ARROW_STUB);
+  const entryX = Math.max(0, target.x - DEPENDENCY_ARROW_TARGET_CLEARANCE);
+  return [
+    [
+      { x: exitX - DEPENDENCY_ARROW_STUB, y: source.y },
+      { x: exitX, y: source.y },
+      { x: exitX, y: detourY },
+      { x: entryX, y: detourY },
+      { x: entryX, y: target.y },
+      { x: entryX + DEPENDENCY_ARROW_TARGET_CLEARANCE, y: target.y },
+    ],
+  ];
+}
+
+/**
+ * How much a route segment must overlap an event to count as crossing it — anchors
+ * touching their own event's edge must not count.
+ */
+const COLLISION_EPSILON = 0.5;
+
+function segmentCrossesObstacle(
+  a: DependencyArrowPoint,
+  b: DependencyArrowPoint,
+  obstacle: DependencyArrowObstacle,
+): boolean {
+  // Vertical segment: x strictly inside the obstacle, y ranges overlapping.
+  if (a.x === b.x) {
+    return (
+      a.x > obstacle.x1 + COLLISION_EPSILON &&
+      a.x < obstacle.x2 - COLLISION_EPSILON &&
+      Math.min(Math.max(a.y, b.y), obstacle.y2) - Math.max(Math.min(a.y, b.y), obstacle.y1) >
+        COLLISION_EPSILON
+    );
+  }
+  // Horizontal segment: y strictly inside the obstacle, x ranges overlapping.
+  if (a.y === b.y) {
+    return (
+      a.y > obstacle.y1 + COLLISION_EPSILON &&
+      a.y < obstacle.y2 - COLLISION_EPSILON &&
+      Math.min(Math.max(a.x, b.x), obstacle.x2) - Math.max(Math.min(a.x, b.x), obstacle.x1) >
+        COLLISION_EPSILON
+    );
+  }
+  // Routes only have horizontal and vertical segments.
+  return false;
+}
+
+/**
+ * Number of obstacles a route crosses. Each obstacle counts once, no matter how many
+ * segments cross it.
+ */
+export function countRouteCollisions(
+  points: readonly DependencyArrowPoint[],
+  obstacles: readonly DependencyArrowObstacle[],
+): number {
+  let count = 0;
+  for (const obstacle of obstacles) {
+    for (let index = 0; index < points.length - 1; index += 1) {
+      if (segmentCrossesObstacle(points[index], points[index + 1], obstacle)) {
+        count += 1;
+        break;
+      }
+    }
+  }
+  return count;
+}
+
+function formatCoordinate(value: number): number {
+  return Math.round(value * 100) / 100;
+}
+
+/**
+ * Builds an SVG path following the provided points with horizontal/vertical segments,
+ * softening each corner with a quadratic curve. The radius is clamped so two corners
+ * never overlap, and zero-length segments collapse their corner.
+ */
+export function buildRoundedOrthogonalPath(
+  points: readonly DependencyArrowPoint[],
+  radius: number,
+): string {
+  // Drop consecutive duplicated points so degenerate segments don't produce corners.
+  const cleanPoints = points.filter(
+    (point, index) =>
+      index === 0 || point.x !== points[index - 1].x || point.y !== points[index - 1].y,
+  );
+
+  if (cleanPoints.length < 2) {
+    return '';
+  }
+
+  let d = `M ${formatCoordinate(cleanPoints[0].x)} ${formatCoordinate(cleanPoints[0].y)}`;
+
+  for (let index = 1; index < cleanPoints.length - 1; index += 1) {
+    const previous = cleanPoints[index - 1];
+    const corner = cleanPoints[index];
+    const next = cleanPoints[index + 1];
+
+    const inLength = Math.hypot(corner.x - previous.x, corner.y - previous.y);
+    const outLength = Math.hypot(next.x - corner.x, next.y - corner.y);
+    const cornerRadius = Math.min(radius, inLength / 2, outLength / 2);
+
+    if (cornerRadius <= 0) {
+      d += ` L ${formatCoordinate(corner.x)} ${formatCoordinate(corner.y)}`;
+      continue;
+    }
+
+    const inX = corner.x - ((corner.x - previous.x) / inLength) * cornerRadius;
+    const inY = corner.y - ((corner.y - previous.y) / inLength) * cornerRadius;
+    const outX = corner.x + ((next.x - corner.x) / outLength) * cornerRadius;
+    const outY = corner.y + ((next.y - corner.y) / outLength) * cornerRadius;
+
+    d += ` L ${formatCoordinate(inX)} ${formatCoordinate(inY)}`;
+    d += ` Q ${formatCoordinate(corner.x)} ${formatCoordinate(corner.y)} ${formatCoordinate(outX)} ${formatCoordinate(outY)}`;
+  }
+
+  const last = cleanPoints[cleanPoints.length - 1];
+  d += ` L ${formatCoordinate(last.x)} ${formatCoordinate(last.y)}`;
+
+  return d;
+}
